requirements.txt
dspy-ai
google-generativeai
python-dotenv # For potentially loading .env files if you choose to use them
polars
requests
transformers
torch
openpyxl
